move entities to db
